golem_base_sdk

Module eth

Module for Ethereum transaction-related functionality. Provides helpers for constructing, signing, and sending Ethereum transactions.

Structs

GolemBaseStorageEntityBTLExtended

Event with signature GolemBaseStorageEntityBTLExtended(uint256,uint256,uint256) and selector 0x835bfca6df78ffac92635dcc105a6a8c4fd715e054e18ef60448b0a6dce30c8d.

GolemBaseStorageEntityCreated

Event with signature GolemBaseStorageEntityCreated(uint256,uint256) and selector 0xce4b4ad6891d716d0b1fba2b4aeb05ec20edadb01df512263d0dde423736bbb9.

GolemBaseStorageEntityDeleted

Event with signature GolemBaseStorageEntityDeleted(uint256) and selector 0x0297b0e6eaf1bc2289906a8123b8ff5b19e568a60d002d47df44f8294422af93.

GolemBaseStorageEntityUpdated

Event with signature GolemBaseStorageEntityUpdated(uint256,uint256) and selector 0xf371f40aa6932ad9dacbee236e5f3b93d478afe3934b5cfec5ea0d800a41d165.

Enums

Error

Represents errors that can occur in the GolemBase ETH client. Used for wrapping transaction, receipt, and log decoding errors.

Constants

STORAGE_ADDRESS

The Ethereum address of the GolemBase storage contract. All entity-related transactions are sent to this address.