golem_base_sdk::signers
Trait TransactionSigner

pub trait TransactionSigner: Send + Sync {
    // Required methods
    fn address(&self) -> Address;
    fn sign<'life0, 'life1, 'async_trait>(
        &'life0 self,
        data: &'life1 [u8],
    ) -> Pin<Box<dyn Future<Output = Result<Signature>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
}

Trait for signing transactions with different backends. Implementors provide an address and a method to sign arbitrary data.

Required Methods

fn address(&self) -> Address

Returns the address of the signer as an Address.

fn sign<'life0, 'life1, 'async_trait>( &'life0 self, data: &'life1 [u8], ) -> Pin<Box<dyn Future<Output = Result<Signature>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Signs the given data and returns a Signature.