golem_base_sdk

Module eth

Module for Ethereum transaction-related functionality. Provides helpers for constructing, signing, and sending Ethereum transactions.

Modules

GolemBaseABI

Module containing a contract’s types and functions.

Enums

Error

Represents errors that can occur in the GolemBase ETH client. Used for wrapping transaction, receipt, and log decoding errors.

Constants

STORAGE_ADDRESS

The Ethereum address of the GolemBase storage contract. All entity-related transactions are sent to this address.